env_check

Verify that object names in the environment defined by the pos parameter are identical or not to object names in the above environment (following R Scope). This can be used to verify that names used for objects inside a function or in the working environment do not override names of objects already present in the above R environments, following the R scope.

env_check(pos = 1, name = NULL)

Arguments

pos

Single non nul positive integer indicating the position of the environment checked (argument n of the parent.frame() function). Value 1 means one step above the env_check() local environment (by default). This means that when env_check(pos = 1) is used inside a function A, it checks if the name of any object in the local environment of this function A is also present in above environments, following R Scope, starting by the just above environment. When env_check(pos = 1) is used in the working (Global) environment (named .GlobalEnv), it checks the object names of this .GlobalEnv environment, in the above environments. See the examples below.

name

Single character string indicating a string that will be added in the output string, for instance the name of a function inside which env_check() is used.

Value

A character string indicating the object names of the tested environment that match object names in the above environments, following the R scope, or NULL if no match.

Details

REQUIRED PACKAGES

None

REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

arg_check()

Examples

# Example in the working environment

mean <- 1 
# creation of the object mean with value 1 in the .GlobalEnv environment, 
# knowing that the mean() function also exists in the environment base, above .GlobalEnv.
t.test <- 1 
# creation of the object t.test with value 1 in the .GlobalEnv environment, 
# knowing that the t.test() function also exists in the environment stats, above .GlobalEnv.
search() 
#>  [1] ".GlobalEnv"        "package:cuteDev"   "package:stats"    
#>  [4] "package:graphics"  "package:grDevices" "package:utils"    
#>  [7] "package:datasets"  "package:methods"   "Autoloads"        
#> [10] "tools:callr"       "package:base"     
# current R scope (order of the successive R environments).
utils::find("mean") 
#> [1] "package:base"
# where the objects with the name "mean" are present.
utils::find("t.test") 
#> [1] "package:stats"
# where the objects with the name "mean" are present.
a <- env_check(pos = 1) 
# test if any object name of the global environment are above environments 
# (or env_check(), as pos = 1 is the default value).
a # the output string of sec().
#> [1] "SOME VARIABLES OF THE CHECKED ENVIRONMENT ARE ALSO PRESENT IN :\npackage:stats: t.test\npackage:base: mean\n"
cat(a) # the evaluated output.
#> SOME VARIABLES OF THE CHECKED ENVIRONMENT ARE ALSO PRESENT IN :
#> package:stats: t.test
#> package:base: mean
env_check(pos = 2) 
#> [1] "SOME VARIABLES OF THE CHECKED ENVIRONMENT ARE ALSO PRESENT IN :\neval_with_user_handlers(expr, envir, enclos, user_handlers): enclos envir expr\nwithCallingHandlers(withVisible(eval_with_user_handlers(expr, envir, enclos, user_handlers)), warning = wHandler, error = eHandler, message = mHandler): expr\ndoTryCatch(return(expr), name, parentenv, handler): expr\ntryCatchOne(expr, names, parentenv, handlers[[1]]): expr\ntryCatchList(expr, classes, parentenv, handlers): expr\ntryCatch(expr, error = function(e) {\n    call <- conditionCall(e)\n    if (!is.null(call)) {\n        if (identical(call[[1]], quote(doTryCatch))) \n            call <- sys.call(-4)\n        dcall <- deparse(call, nlines = 1)\n        prefix <- paste(\"Error in\", dcall, \": \")\n        LONG <- 75\n        sm <- strsplit(conditionMessage(e), \"\\n\")[[1]]\n        w <- 14 + nchar(dcall, type = \"w\") + nchar(sm[1], type = \"w\")\n        if (is.na(w)) \n            w <- 14 + nchar(dcall, type = \"b\") + nchar(sm[1], type = \"b\")\n        if (w > LONG) \n            prefix <- paste0(prefix, \"\\n  \")\n    }\n    else prefix <- \"Error : \"\n    msg <- paste0(prefix, conditionMessage(e), \"\\n\")\n    .Internal(seterrmessage(msg[1]))\n    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {\n        cat(msg, file = outFile)\n        .Internal(printDeferredWarnings())\n    }\n    invisible(structure(msg, class = \"try-error\", condition = e))\n}): expr\ntry(f, silent = TRUE): expr\nevaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, debug = debug, last = i == length(out), use_try = stop_on_error != 2, keep_warning = keep_warning, keep_message = keep_message, log_echo = log_echo, log_warning = log_warning, output_handler = output_handler, include_timing = include_timing): enclos envir expr\nevaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler): enclos envir expr\nwithCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env, run_dont_run = run_dont_run), error = function(err) {\n    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), i = err$message)\n    abort(msg, parent = err)\n}): expr\nwithCallingHandlers(expr, error = function(cnd) {\n    if (i == 0) {\n    }\n    else {\n        message <- c(i = \"In index: {i}.\")\n        if (!is.null(names) && !is.na(names[[i]]) && names[[i]] != \"\") {\n            name <- names[[i]]\n            message <- c(message, i = \"With name: {name}.\")\n        }\n        else {\n            name <- NULL\n        }\n        cli::cli_abort(message, location = i, name = name, parent = cnd, call = error_call, class = \"purrr_error_indexed\")\n    }\n}): expr\nwith_indexed_errors(i = i, names = names, error_call = .purrr_error_call, call_with_cleanup(map_impl, environment(), .type, .progress, n, names, i)): expr\n(function (what, args, quote = FALSE, envir = parent.frame()) \n{\n    if (!is.list(args)) \n        stop(\"second argument must be a list\")\n    if (quote) \n        args <- lapply(args, enquote)\n    .Internal(do.call(what, args, envir))\n})(base::quote(function (..., cli_colors, pkgdown_internet) \n{\n    options(cli.num_colors = cli_colors, crayon.colors = cli_colors, pkgdown.internet = pkgdown_internet)\n    pkgdown::build_site(...)\n}), base::quote(list(pkg = list(package = \"cuteDev\", version = \"1.0\", lang = \"en\", src_path = \"C:/Users/yhan/Documents/Git_projects/cuteDev\", dst_path = \"C:/Users/yhan/Documents/Git_projects/cuteDev/docs\", install_metadata = FALSE, desc = <environment>, meta = list(), figures = list(dev = \"ragg::agg_png\", dpi = 96, dev.args = list(), fig.ext = \"png\", fig.width = 7.29166666666667, fig.height = NULL, fig.retina = 2, fig.asp = 0.618046971569839, bg = NULL, other.parameters = list()), repo = NULL, development = list(\n    destination = \"dev\", mode = \"default\", version_label = \"default\", in_dev = FALSE), topics = list(name = c(arg_check.Rd = \"arg_check\", arg_test.Rd = \"arg_test\", env_check.Rd = \"env_check\", get_message.Rd = \"get_message\", internal_function.Rd = \".pack_and_function_check\", is_function_here.Rd = \"is_function_here\", is_package_here.Rd = \"is_package_here\", python_pkg_check.Rd = \"python_pkg_check\", report.Rd = \"report\"), file_in = c(\"arg_check.Rd\", \"arg_test.Rd\", \"env_check.Rd\", \"get_message.Rd\", \"internal_function.Rd\", \n\"is_function_here.Rd\", \"is_package_here.Rd\", \"python_pkg_check.Rd\", \"report.Rd\"), file_out = c(\"arg_check.html\", \"arg_test.html\", \"env_check.html\", \"get_message.html\", \"internal_function.html\", \"is_function_here.html\", \"is_package_here.html\", \"python_pkg_check.html\", \"report.html\"), alias = list(arg_check.Rd = \"arg_check\", arg_test.Rd = \"arg_test\", env_check.Rd = \"env_check\", get_message.Rd = \"get_message\", internal_function.Rd = \".pack_and_function_check\", is_function_here.Rd = \"is_function_here\", \n    is_package_here.Rd = \"is_package_here\", python_pkg_check.Rd = \"python_pkg_check\", report.Rd = \"report\"), funs = list(arg_check.Rd = \"arg_check()\", arg_test.Rd = \"arg_test()\", env_check.Rd = \"env_check()\", get_message.Rd = \"get_message()\", internal_function.Rd = \".pack_and_function_check()\", is_function_here.Rd = \"is_function_here()\", is_package_here.Rd = \"is_package_here()\", python_pkg_check.Rd = \"python_pkg_check()\", report.Rd = \"report()\"), title = c(arg_check.Rd = \"arg_check\", arg_test.Rd = \"arg_test\", \nenv_check.Rd = \"env_check\", get_message.Rd = \"get_message\", internal_function.Rd = \".pack_and_function_check\", is_function_here.Rd = \"is_function_here\", is_package_here.Rd = \"is_package_here\", python_pkg_check.Rd = \"python_pkg_check\", report.Rd = \"report\"), rd = list(arg_check.Rd = list(\"% Generated by roxygen2: do not edit by hand\", \"\\n\", \"% Please edit documentation in R/arg_check.R\", \"\\n\", list(\"arg_check\"), \"\\n\", list(\"arg_check\"), \"\\n\", list(\"arg_check\"), \"\\n\", list(\"\\n\", \"arg_check(\\n\", \"  data,\\n\", \n    \"  class = NULL,\\n\", \"  typeof = NULL,\\n\", \"  mode = NULL,\\n\", \"  length = NULL,\\n\", \"  prop = FALSE,\\n\", \"  double.as.integer.allowed = FALSE,\\n\", \"  options = NULL,\\n\", \"  all.options.in.data = FALSE,\\n\", \"  na.contain = FALSE,\\n\", \"  neg.values = TRUE,\\n\", \"  inf.values = TRUE,\\n\", \"  print = FALSE,\\n\", \"  data.name = NULL,\\n\", \"  fun.name = NULL\\n\", \")\\n\"), \"\\n\", list(\"\\n\", list(list(\"data\"), list(\"Object to test.\")), \"\\n\", \"\\n\", list(list(\"class\"), list(\"Single character string. Either one of the class() result (But see the warning section above) or \\\"vector\\\" or \\\"ggplot2\\\" (i.e., objects of class c(\\\"gg\\\", \\\"ggplot\\\")) or NULL.\")), \n    \"\\n\", \"\\n\", list(list(\"typeof\"), list(\"Single character string. Either one of the typeof() result or NULL.\")), \"\\n\", \"\\n\", list(list(\"mode\"), list(\"Single character string. Either one of the mode() result (for non-vector object) or NULL.\")), \"\\n\", \"\\n\", list(list(\"length\"), list(\"Single numeric value indicating the length of the object. Not considered if NULL.\")), \"\\n\", \"\\n\", list(list(\"prop\"), list(\"Single logical value. Are the numeric values between 0 and 1 (proportion)? If TRUE, can be used alone, without considering class, etc.\")), \n    \"\\n\", \"\\n\", list(list(\"double.as.integer.allowed\"), list(\"Single logical value. If TRUE, no error is reported in the cheking message if argument is set to typeof == \\\"integer\\\" or class == \\\"integer\\\", while the reality is typeof == \\\"double\\\" or class == \\\"numeric\\\" but the numbers strictly have zero as modulo (remainder of a division). This means that i <- 1, which is typeof(i) -> \\\"double\\\" is considered as integer with double.as.integer.allowed = TRUE. WARNING: data mod 1 == 0L but not isTRUE(all.equal(data mod 1, 0)) is used here because the argument checks for integers stored as double (does not check for decimal numbers that are approximate integers).\")), \n    \"\\n\", \"\\n\", list(list(\"options\"), list(\"Vector of character strings or integers indicating all the possible option values for the data argument, or NULL. Numbers of type \\\"double\\\" are accepted if they have a 0 modulo.\")), \"\\n\", \"\\n\", list(list(\"all.options.in.data\"), list(\"Single logical value. If TRUE, all of the options must be present at least once in the data argument, and nothing else. If FALSE, some or all of the options must be present in the data argument, and nothing else. Ignored if options is NULL.\")), \n    \"\\n\", \"\\n\", list(list(\"na.contain\"), list(\"Single logical value. Can the data argument contain NA?\")), \"\\n\", \"\\n\", list(list(\"neg.values\"), list(\"Single logical value. Are negative numeric values authorized? Warning: the default setting is TRUE, meaning that, in that case, no check is performed for the presence of negative values. The neg.values argument is activated only when set to FALSE. In addition, (1) neg.values = FALSE can only be used when class, typeof or mode arguments are not NULL, otherwise return an error message, (2) the presence of negative values is not checked with neg.values = FALSE if the tested object is a factor and the following checking message is returned \\\"OBJECT MUST BE MADE OF NON NEGATIVE VALUES BUT IS A FACTOR\\\".\")), \n    \"\\n\", \"\\n\", list(list(\"inf.values\"), list(\"Single logical value. Are infinite numeric values authorized (Inf or -Inf)? Identical remarks as for the neg.values argument.\")), \"\\n\", \"\\n\", list(list(\"print\"), list(\"Single logical value. Print the message if $problem is TRUE? Warning: set by default to FALSE, which facilitates the control of the checking message output when using arg_check() inside functions. See the example section.\")), \"\\n\", \"\\n\", list(list(\"data.name\"), list(\"Single character string indicating the name of the object to test. If NULL, use what is assigned to the data argument for the returned message.\")), \n    \"\\n\", \"\\n\", list(list(\"fun.name\"), list(\"Single character string indicating the name of the function checked (i.e., when arg_check() is used to check the arguments of this function). If non-null, the value of fun.name will be added into the message returned by arg_check().\")), \"\\n\"), \"\\n\", list(\"\\n\", \"A list containing:\\n\", \"- problem: logical. Is there any problem detected?\\n\", \"- text: message indicating the details of the problem, or the absence of problem.\\n\", \"- object.name: value of the data.name argument (i.e., name of the checked object if provided, NULL otherwise).\\n\"), \n    \"\\n\", list(\"\\n\", \"Check the class, type, mode and length of the data argument.\\n\", \"\\n\", \"Mainly used to check the arguments of other functions.\\n\", \"\\n\", \"Check also other kind of data parameters, is it a proportion? Is it type double but numbers without decimal part?\\n\"), \"\\n\", list(\"\\n\", \"- If options == NULL, then at least class or type or mode or length argument must be non-null.\\n\", \" - If options is non-null, then class, type and mode must be NULL, and length can be NULL or specified.\\n\", \n        \" \\n\", \"\\n\", \"WARNINGS\\n\", \"\\n\", \"The function tests what is written in its arguments, even if what is written is incoherent. For instance, arg_check(data = factor(1), class = \\\"factor\\\", mode = \\\"character\\\") will return a problem, whatever the object tested in the data argument, because no object can be class \\\"factor\\\" and mode \\\"character\\\" (factors are class \\\"factor\\\" and mode \\\"numeric\\\"). Of note, length of object of class \\\"environment\\\" is always 0.\\n\", \"\\n\", \"If the tested object is NULL, then the function will always return a checking problem.\\n\", \n        \"\\n\", \"Since R >= 4.0.0, class(matrix()) returns \\\"matrix\\\" \\\"array\\\", and not \\\"matrix\\\" alone as before. However, use argument class = \\\"matrix\\\" to check for matrix object (of class \\\"matrix\\\" \\\"array\\\" in R >= 4.0.0) and use argument class = \\\"array\\\" to check for array object (of class \\\"array\\\" in R >= 4.0.0).\\n\", \"\\n\", \"\\n\", \"REQUIRED PACKAGES\\n\", \"\\n\", \"None\\n\", \"\\n\", \"\\n\", \"REQUIRED FUNCTIONS FROM THE cute PACKAGE\\n\", \"\\n\", \"None\\n\"), \"\\n\", list(\"\\n\", \"test <- matrix(1:3)\\n\", \"# arg_check(data = test, print = TRUE, class = \\\"vector\\\", mode = \\\"numeric\\\")  # commented because this example returns an error\\n\"), \n    \"\\n\"), arg_test.Rd = list(\"% Generated by roxygen2: do not edit by hand\", \"\\n\", \"% Please edit documentation in R/arg_test.R\", \"\\n\", list(\"arg_test\"), \"\\n\", list(\"arg_test\"), \"\\n\", list(\"arg_test\"), \"\\n\", list(\"\\n\", \"arg_test(\\n\", \"  fun,\\n\", \"  arg,\\n\", \"  val,\\n\", \"  expect.error = NULL,\\n\", \"  parall = FALSE,\\n\", \"  thread.nb = NULL,\\n\", \"  print.count = 10,\\n\", \"  plot.fun = FALSE,\\n\", \"  export = FALSE,\\n\", \"  res.path = NULL,\\n\", \"  lib.path = NULL\\n\", \")\\n\"), \"\\n\", list(\"\\n\", list(list(\n    \"fun\"), list(\"Single character string indicating the name of the function tested (without brackets).\")), \"\\n\", \"\\n\", list(list(\"arg\"), list(\"Vector of character strings of arguments of fun. At least arguments that do not have default values must be present in this vector.\")), \"\\n\", \"\\n\", list(list(\"val\"), list(\"List with number of compartments equal to the length of arg, each compartment containing values of the corresponding argument in arg. Each different value must be in a list or in a vector. For instance, argument 3 in arg is a logical argument (values accepted TRUE, FALSE, NA). Thus, compartment 3 of val can be either list(TRUE, FALSE, NA), or c(TRUE, FALSE, NA). NULL value alone must be written list(NULL).\")), \n    \"\\n\", \"\\n\", list(list(\"expect.error\"), list(\"List of exactly the same structure as val argument, but containing FALSE or TRUE, depending on whether error is expected (TRUE) or not (FALSE) for each corresponding value of val. A message is returned depending on discrepancies between the expected and observed errors. See the examples below. BEWARE: not always possible to write the expected errors for all the combination of argument values. Ignored if NULL.\")), \"\\n\", \"\\n\", list(list(\"parall\"), list(\n        \"Single logical value. Force parallelization ?\")), \"\\n\", \"\\n\", list(list(\"thread.nb\"), list(\"Single numeric integer indicating the number of threads to use if ever parallelization is required. If NULL, all the available threads will be used. Ignored if parall is FALSE.\")), \"\\n\", \"\\n\", list(list(\"print.count\"), list(\"Single interger value. Print a working progress message every print.count during loops. BEWARE: can increase substentially the time to complete the process if using a small integer value, like 10 for instance. Use Inf if no loop message desired.\")), \n    \"\\n\", \"\\n\", list(list(\"plot.fun\"), list(\"Single logical value. Plot the plotting function tested for each test? Ignored if the tested function is not a graphic function.\")), \"\\n\", \"\\n\", list(list(\"export\"), list(\"Single logical value. Export the results into a .RData file and into a .tsv file? If FALSE, return a list into the console (see below). BEWARE: will be automatically set to TRUE if parall is TRUE. This means that when using parallelization, the results are systematically exported, not returned into the console.\")), \n    \"\\n\", \"\\n\", list(list(\"res.path\"), list(\"Single character string indicating the absolute pathway of the folder where the txt results and pdfs, containing all the plots, will be saved. Several txt and pdf, one per thread, if parallelization. Ignored if export is FALSE. Must be specified if parall is TRUE or if export is TRUE.\")), \"\\n\", \"\\n\", list(list(\"lib.path\"), list(\"Vector of characters specifying the absolute pathways of the directories containing the required packages if not in the default directories. Ignored if NULL.\")), \n    \"\\n\"), \"\\n\", list(\"\\n\", \"One or several pdf if a plotting function is tested and if the plot.fun argument is TRUE. \\n\", \"\\n\", \"And then, if export is FALSE a list containing:\\n\", \"- $fun: the tested function.\\n\", \"- $ini: the initial input values.\\n\", \"- $data: a data frame of all the combination tested, containing the following columns:\\n\", \"     # the different values tested, named by arguments\\n\", \"     - $kind: a vector of character strings indicating the kind of test result: either \\\"ERROR\\\", or \\\"WARNING\\\", or \\\"OK\\\".\\n\", \n    \"     - $problem: a logical vector indicating if error or not.\\n\", \"     - $expected.error: optional logical vector indicating the expected error specified in the expect.error argument.\\n\", \"     - $message: either NULL if $kind is always \\\"OK\\\", or the messages.\\n\", \"- $sys.info: system and packages info.\\n\", \"\\n\", \"If export is TRUE:\\n\", \"- the same list object into a .RData file.\\n\", \"- also the $data data frame into a .tsv file.\\n\", \"- if expect.error is non NULL and if any discrepancy, the $data data frame into a .tsv file but containing only the rows with discrepancies between expected and observed errors.\\n\"), \n    \"\\n\", list(\"\\n\", \"Test combinations of argument values of a function.\\n\"), \"\\n\", list(\"\\n\", \"REQUIRED PACKAGES\\n\", \"\\n\", \"lubridate\\n\", \"\\n\", \"parallel if parall argument is TRUE (included in the R installation packages but not automatically loaded)\\n\", \"\\n\", \"pdftools if parall argument is TRUE (included in the R installation packages but not automatically loaded)\\n\", \"\\n\", \"If the tested function is in a package, this package must be imported first (no parallelization) or must be in the classical R package folder indicated by the lib.path argument (parallelization)\\n\", \n        \"\\n\", \"\\n\", \"REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION\\n\", \"\\n\", \"arg_check()\\n\", \"\\n\", \"get_message()\\n\", \"\\n\", \"WARNINGS\\n\", \"\\n\", \"Limited to 43 arguments with at least 2 values each. The total number of arguments tested can be more if the additional arguments have a single value. The limit is due to nested \\\"for\\\" loops (https://stat.ethz.ch/pipermail/r-help/2008-March/157341.html), but this limitation is away from the number of tests performed that would be 2^43.\\n\"), \"\\n\", list(\n        \"\\n\", \"arg_test(fun = \\\"unique\\\", arg = c(\\\"x\\\", \\\"incomparables\\\"), \\n\", \"val = list(x = list(1:10, c(1,1,2,8), NA), incomparable = c(TRUE, FALSE, NA)))\\n\", \"\\n\", \"arg_test(fun = \\\"unique\\\", arg = c(\\\"x\\\", \\\"incomparables\\\"), \\n\", \"val = list(x = list(1:10, c(1,1,2,8), NA), \\n\", \"incomparable = c(TRUE, FALSE, NA)), expect.error = list(x = list(FALSE, FALSE, TRUE), \\n\", \"incomparable = c(FALSE, FALSE, TRUE)))\\n\", \"\\n\", \"arg_test(fun = \\\"unique\\\", arg = c(\\\"x\\\", \\\"incomparables\\\"), \\n\", \"val = list(x = list(1:10, c(1,1,2,8), NA), incomparable = c(TRUE, FALSE, NA)), \\n\", \n        \"expect.error = list(x = list(FALSE, FALSE, TRUE), incomparable = c(FALSE, FALSE, TRUE)), \\n\", \"export = TRUE, res.path = getwd())\\n\", \"\\n\", \"# Return an error if you run this example\\n\", \"# arg_test(fun = \\\"round\\\", arg = c(\\\"data\\\", \\\"dec.nb\\\", \\\"after.lead.zero\\\"), val = list(L1 = list(c(1, 1.0002256, 1.23568), \\\"a\\\", NA), L2 = list(2, c(1,3), NA), L3 = c(TRUE, FALSE, NA)))\\n\", \"\\n\", \"arg_test(fun = \\\"plot\\\", arg = c(\\\"x\\\", \\\"y\\\"), val = list(x = list(1:10, 12:13, NA, (1:10)^2), \\n\", \"y = list(1:10, NA, NA)),  expect.error = list(x = list(FALSE, TRUE, TRUE, FALSE), \\n\", \n        \"y = list(FALSE, TRUE, TRUE)), parall = FALSE, thread.nb = NULL, plot.fun = TRUE, \\n\", \"res.path = \\\".\\\", lib.path = NULL)\\n\", \"\\n\", \"arg_test(fun = \\\"plot\\\", arg = c(\\\"x\\\", \\\"y\\\"), val = list(x = list(1:10, 12:13, NA, \\n\", \"(1:10)^2), y = list(1:10, NA, NA)), parall = FALSE, thread.nb = 4, \\n\", \"plot.fun = TRUE, res.path = \\\".\\\", \\n\", \"lib.path = NULL)\\n\", \"\\n\", \"# set.seed(1) ; \\n\", \"# obs1 <- data.frame(Time = c(rnorm(10), rnorm(10) + 2), Group1 = rep(c(\\\"G\\\", \\\"H\\\"), each = 10), \\n\", \n        \"# stringsAsFactors = TRUE) ; arg_test(fun = \\\"fun_gg_boxplot\\\", arg = c(\\\"data1\\\", \\\"y\\\", \\\"categ\\\"), \\n\", \"# val = list(L1 = list(L1 = obs1), L2 = list(L1 = \\\"Time\\\"), L3 = list(L1 = \\\"Group1\\\")))\\n\", \"\\n\", \"# set.seed(1) ; \\n\", \"# obs1 <- data.frame(Time = c(rnorm(10), rnorm(10) + 2), Group1 = rep(c(\\\"G\\\", \\\"H\\\"), each = 10), \\n\", \"# stringsAsFactors = TRUE) ; \\n\", \"# arg_test(fun = \\\"fun_gg_boxplot\\\", arg = c(\\\"data1\\\", \\\"y\\\", \\\"categ\\\"), val = list(L1 = list(obs1), \\n\", \"# L2 = \\\"Time\\\", L3 = \\\"Group1\\\"), parall = FALSE, thread.nb = NULL, plot.fun = TRUE, \\n\", \n        \"# res.path = \\\"C:\\\\Users\\\\yhan\\\\Desktop\\\\\\\", lib.path = \\\"C:\\\\Program Files\\\\R\\\\R-4.3.1\\\\library\\\\\\\")\\n\", \"\\n\", \"# library(ggplot2) ; arg_test(fun = \\\"geom_histogram\\\", arg = c(\\\"data\\\", \\\"mapping\\\"), \\n\", \"# val = list(x = list(data.frame(X = \\\"a\\\", stringsAsFactors = TRUE)), \\n\", \"# y = list(ggplot2::aes(x = X))), parall = FALSE, thread.nb = NULL, \\n\", \"# plot.fun = TRUE, res.path = \\\"C:\\\\Users\\\\yhan\\\\Desktop\\\\\\\", \\n\", \"# lib.path = \\\"C:\\\\Program Files\\\\R\\\\R-4.3.1\\\\library\\\\\\\") \\n\", \"# BEWARE: ggplot2::geom_histogram does not work\\n\"), \n    \"\\n\"), env_check.Rd = list(\"% Generated by roxygen2: do not edit by hand\", \"\\n\", \"% Please edit documentation in R/env_check.R\", \"\\n\", list(\"env_check\"), \"\\n\", list(\"env_check\"), \"\\n\", list(\"env_check\"), \"\\n\", list(\"\\n\", \"env_check(pos = 1, name = NULL)\\n\"), \"\\n\", list(\"\\n\", list(list(\"pos\"), list(\"Single non nul positive integer indicating the position of the environment checked (argument n of the parent.frame() function). Value 1 means one step above the env_check() local environment (by default). This means that when env_check(pos = 1) is used inside a function A, it checks if the name of any object in the local environment of this function A is also present in above environments, following R Scope, starting by the just above environment. When env_check(pos = 1) is used in the working (Global) environment (named .GlobalEnv), it checks the object names of this .GlobalEnv environment, in the above environments. See the examples below.\")), \n    \"\\n\", \"\\n\", list(list(\"name\"), list(\"Single character string indicating a string that will be added in the output string, for instance the name of a function inside which env_check() is used.\")), \"\\n\"), \"\\n\", list(\"\\n\", \"A character string indicating the object names of the tested environment that match object names in the above environments, following the R scope, or NULL if no match.\\n\"), \"\\n\", list(\"\\n\", \"Verify that object names in the environment defined by the pos parameter are identical or not to object names in the above environment (following R Scope). This can be used to verify that names used for objects inside a function or in the working environment do not override names of objects already present in the above R environments, following the R scope.\\n\"), \n    \"\\n\", list(\"\\n\", \"REQUIRED PACKAGES\\n\", \"\\n\", \"None\\n\", \"\\n\", \"\\n\", \"REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION\\n\", \"\\n\", \"arg_check()\\n\"), \"\\n\", list(\"\\n\", \"# Example in the working environment\\n\", \"\\n\", \"mean <- 1 \\n\", \"# creation of the object mean with value 1 in the .GlobalEnv environment, \\n\", \"# knowing that the mean() function also exists in the environment base, above .GlobalEnv.\\n\", \"t.test <- 1 \\n\", \"# creation of the object t.test with value 1 in the .GlobalEnv environment, \\n\", \n        \"# knowing that the t.test() function also exists in the environment stats, above .GlobalEnv.\\n\", \"search() \\n\", \"# current R scope (order of the successive R environments).\\n\", \"utils::find(\\\"mean\\\") \\n\", \"# where the objects with the name \\\"mean\\\" are present.\\n\", \"utils::find(\\\"t.test\\\") \\n\", \"# where the objects with the name \\\"mean\\\" are present.\\n\", \"a <- env_check(pos = 1) \\n\", \"# test if any object name of the global environment are above environments \\n\", \"# (or env_check(), as pos = 1 is the default value).\\n\", \n        \"a # the output string of sec().\\n\", \"cat(a) # the evaluated output.\\n\", \"env_check(pos = 2) \\n\", \"# test if any object of the stats environment (one step above .GlobalEnv) \\n\", \"# are above environments. Returns NULL since no object names of stats are in above environments\\n\", \"\\n\", \"\\n\", \"# Example inside a function\\n\", \"\\n\", \"fun1 <- function(){t.test <- 0 ; mean <- 5 ; env_check(pos = 1)} \\n\", \"# env_check() will check if the object names inside the fun1 function \\n\", \"# exist in the .GlobalEnv environment and above.\\n\", \n        \"fun1()\\n\", \"fun2 <- function(){t.test <- 0 ; mean <- 5 ; env_check(pos = 2)} \\n\", \"# env_check() will check if the object names inside the fun2 function \\n\", \"# exist in the stats environment and above.\\n\", \"fun2()\\n\", \"fun3 <- function(){t.test <- 0 ; mean <- 5 ; env_check(pos = 2, name = \\\"fun3\\\")} \\n\", \"# idem fun2() but with the name of the function fun2 indicated. \\n\", \"# Instead of writting name = \\\"fun3\\\", \\n\", \"# we can also use name = as.character(sys.calls()[[length(sys.calls())]]), \\n\", \n        \"# as sys.calls() gives the function name at top stack of the imbricated functions, \\n\", \"# sys.calls()[[length(sys.calls())]] the name of the just above function. \\n\", \"# This can also been used for the above function: as.character(sys.call(1))\\n\", \"fun3()\\n\", \"test.pos <- 1\\n\", \"env_check(pos = test.pos, \\n\", \"name = if(length(sys.calls()) >= test.pos)\\n\", \"{as.character(sys.calls()[[length(sys.calls()) + 1 - test.pos]])}\\n\", \"else{search()[ (1:length(search()))[test.pos - length(sys.calls())]]}) \\n\", \n        \"# here is a way to have the name of the tested environment according to test.pos value\\n\"), \"\\n\"), get_message.Rd = list(\"% Generated by roxygen2: do not edit by hand\", \"\\n\", \"% Please edit documentation in R/get_message.R\", \"\\n\", list(\"get_message\"), \"\\n\", list(\"get_message\"), \"\\n\", list(\"get_message\"), \"\\n\", list(\"\\n\", \"get_message(\\n\", \"  data,\\n\", \"  kind = \\\"error\\\",\\n\", \"  header = TRUE,\\n\", \"  print.no = FALSE,\\n\", \"  text = NULL,\\n\", \"  env = NULL\\n\", \")\\n\"), \"\\n\", list(\"\\n\", list(\n    list(\"data\"), list(\"Single character string to evaluate.\")), \"\\n\", \"\\n\", list(list(\"kind\"), list(\"Single character string. Either \\\"error\\\" to get error messages, or \\\"warning\\\" to get warning messages, or \\\"message\\\" to get non error and non warning messages.\")), \"\\n\", \"\\n\", list(list(\"header\"), list(\"Single logical value. Add a header in the returned message?\")), \"\\n\", \"\\n\", list(list(\"print.no\"), list(\"Single logical value. Print a message saying that no message reported?\")), \"\\n\", \"\\n\", list(\n    list(\"text\"), list(\"Single character string added to the output message (even if no message exists and print.no is TRUE). Inactivated if the header argument is FALSE. Write NULL if not required.\")), \"\\n\", \"\\n\", list(list(\"env\"), list(\"The name of an existing environment. Write NULL if not required.\")), \"\\n\"), \"\\n\", list(\"\\n\", \"The message or NULL if no message and print.no is FALSE.\\n\"), \"\\n\", list(\"\\n\", \"Evaluate an instruction written between \\\"\\\" and return the first of the error, or warning or standard (non error non warning) messages if ever exist.\\n\", \n    \"\\n\", \"Using argument print.no = FALSE, return NULL if no message, which is convenient in some cases.\\n\"), \"\\n\", list(\"\\n\", \"REQUIRED PACKAGES\\n\", \"\\n\", \"ggplot2\\n\", \"\\n\", \"\\n\", \"REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION\\n\", \"\\n\", \"arg_check()\\n\", \"\\n\", \"WARNINGS\\n\", \"\\n\", \"Only the first message is returned.\\n\", \"\\n\", \"Always use the env argument when get_message() is used inside functions.\\n\", \"\\n\", \"The function does not prevent printing if print() is used inside the instruction tested. To prevent that, use tempo <- utils::capture.output(error <- get_message(data = \\\"arg_check(data = 'a', class = mean, neg.values = FALSE, print = TRUE)\\\")). The return of get_message() is assigned into error and the printed messages are captured by utils::capture.output() and assigned into tempo. See the examples.\\n\"), \n    \"\\n\", list(\"\\n\", \"get_message(data = \\\"wilcox.test(c(1,1,3), c(1, 2, 4), paired = TRUE)\\\", kind = \\\"error\\\", \\n\", \"print.no = TRUE, text = \\\"IN A\\\")\\n\", \"\\n\", \"get_message(data = \\\"wilcox.test(c(1,1,3), c(1, 2, 4), paired = TRUE)\\\", kind = \\\"warning\\\", \\n\", \"print.no = TRUE, text = \\\"IN A\\\")\\n\", \"\\n\", \"get_message(data = \\\"wilcox.test(c(1,1,3), c(1, 2, 4), paired = TRUE)\\\", kind = \\\"message\\\", \\n\", \"print.no = TRUE, text = \\\"IN A\\\")\\n\", \"\\n\", \"get_message(data = \\\"wilcox.test()\\\", kind = \\\"error\\\", print.no = TRUE, text = \\\"IN A\\\")\\n\", \n        \"\\n\", \"get_message(data = \\\"sum(1)\\\", kind = \\\"error\\\", print.no = TRUE, text = \\\"IN A\\\")\\n\", \"\\n\", \"get_message(data = \\\"message('ahah')\\\", kind = \\\"error\\\", print.no = TRUE, text = \\\"IN A\\\")\\n\", \"\\n\", \"get_message(data = \\\"message('ahah')\\\", kind = \\\"message\\\", print.no = TRUE, text = \\\"IN A\\\")\\n\", \"\\n\", \"get_message(data = \\\"ggplot(data = data.frame(X = 1:10, stringsAsFactors = TRUE), \\n\", \"mapping = aes(x = X)) + geom_histogram()\\\", kind = \\\"message\\\", print.no = TRUE, \\n\", \"text = \\\"IN FUNCTION 1\\\")\\n\", \n        \"\\n\", \"set.seed(1) ; \\n\", \"obs1 <- data.frame(Time = c(rnorm(10), rnorm(10) + 2), \\n\", \"Group1 = rep(c(\\\"G\\\", \\\"H\\\"), each = 10), stringsAsFactors = TRUE) ; \\n\", \"get_message(data = 'gg_boxplot(data = obs1, y = \\\"Time\\\", categ = \\\"Group1\\\")', \\n\", \"kind = \\\"message\\\", print.no = TRUE, text = \\\"IN FUNCTION 1\\\")\\n\", \"\\n\"), \"\\n\"), internal_function.Rd = list(\"% Generated by roxygen2: do not edit by hand\", \"\\n\", \"% Please edit documentation in R/internal_functions.R\", \"\\n\", list(\".pack_and_function_check\"), \n    \"\\n\", list(\".pack_and_function_check\"), \"\\n\", list(\".pack_and_function_check\"), \"\\n\", list(\"\\n\", \".pack_and_function_check(fun, lib.path, external.function.name)\\n\"), \"\\n\", list(\"\\n\", list(list(\"fun\"), list(\"Character vector of the names of the required functions, preceded by the name of the package they belong to and a double colon. Example: c(\\\"ggplot2::geom_point\\\", \\\"grid::gpar\\\").\")), \"\\n\", \"\\n\", list(list(\"lib.path\"), list(\"Character vector specifying the absolute pathways of the directories containing the listed packages in the fun argument, if not in the default directories. If NULL, the function checks only in the .libPaths() default R library folders.\")), \n        \"\\n\", \"\\n\", list(list(\"external.function.name\"), list(\"Name of the function using the .pack_and_function_check() function.\")), \"\\n\"), \"\\n\", list(\"\\n\", \"An error message if at least one of the checked packages is missing in lib.path, or if at least one of the checked functions is missing in the required package, nothing otherwise.\\n\"), \"\\n\", list(\"\\n\", \"Check if 1) required functions are present in required packages and 2) required packages are installed locally.\\n\"), \"\\n\", list(\"\\n\", \"# .pack_and_function_check(fun = \\\"ggplot2::notgood\\\") # commented because this example returns an error\\n\", \n        list(\"\\n\", \"# Example that shouldn't be run because this is an internal function\\n\", \".pack_and_function_check(fun = c(\\\"ggplot2::geom_point\\\", \\\"grid::gpar\\\"))\\n\"), \"\\n\"), \"\\n\", list(\"internal\"), \"\\n\"), is_function_here.Rd = list(\"% Generated by roxygen2: do not edit by hand\", \"\\n\", \"% Please edit documentation in R/is_function_here.R\", \"\\n\", list(\"is_function_here\"), \"\\n\", list(\"is_function_here\"), \"\\n\", list(\"is_function_here\"), \"\\n\", list(\"\\n\", \"is_function_here(fun, lib.path = NULL)\\n\"), \n    \"\\n\", list(\"\\n\", list(list(\"fun\"), list(\"Character vector of the names of the required functions, preceded by the name of the package they belong to and a double colon. Example: c(\\\"ggplot2::geom_point\\\", \\\"grid::gpar\\\"). Warning: do not write \\\"()\\\" at the end of the function\")), \"\\n\", \"\\n\", list(list(\"lib.path\"), list(\"Character vector specifying the absolute pathways of the directories containing the listed packages in the fun argument, if not in the default directories. Ignored if NULL.\")), \n        \"\\n\"), \"\\n\", list(\"\\n\", \"An error message if at least one of the checked packages is missing in lib.path, or if at least one of the checked functions is missing in the required package, nothing otherwise.\\n\"), \"\\n\", list(\"\\n\", \"Check if required functions are present in installed packages. This controls a change of function name, for example.\\n\"), \"\\n\", list(\"\\n\", \"# is_function_here(fun = \\\"ggplot2::notgood\\\") # commented because this example returns an error\\n\", \"is_function_here(fun = c(\\\"ggplot2::geom_point\\\", \\\"grid::gpar\\\"))\\n\"), \n    \"\\n\"), is_package_here.Rd = list(\"% Generated by roxygen2: do not edit by hand\", \"\\n\", \"% Please edit documentation in R/is_package_here.R\", \"\\n\", list(\"is_package_here\"), \"\\n\", list(\"is_package_here\"), \"\\n\", list(\"is_package_here\"), \"\\n\", list(\"\\n\", \"is_package_here(req.package, lib.path = NULL)\\n\"), \"\\n\", list(\"\\n\", list(list(\"req.package\"), list(\"Character vector of package names to check.\")), \"\\n\", \"\\n\", list(list(\"lib.path\"), list(\"Character vector specifying the absolute pathways of the directories containing the listed packages in the req.package argument, if not in the default directories. If NULL, the function checks only in the .libPaths() default R library folders.\")), \n    \"\\n\"), \"\\n\", list(\"\\n\", \"An error message if at least one of the checked packages is missing in lib.path, nothing otherwise.\\n\"), \"\\n\", list(\"\\n\", \"Check if required packages are installed locally.\\n\"), \"\\n\", list(\"\\n\", \"# is_package_here(req.package = \\\"nopackage\\\") # commented because this example returns an error\\n\", \"is_package_here(req.package = \\\"ggplot2\\\")\\n\", \"# is_package_here(req.package = \\\"ggplot2\\\", lib.path = \\\"C:/Users/yhan/AppData/Local/R/win-library/4.3\\\") \\n\", \"# commented because this example returns an error if the lib.path argument is not an existing directory\\n\"), \n    \"\\n\"), python_pkg_check.Rd = list(\"% Generated by roxygen2: do not edit by hand\", \"\\n\", \"% Please edit documentation in R/python_pkg_check.R\", \"\\n\", list(\"python_pkg_check\"), \"\\n\", list(\"python_pkg_check\"), \"\\n\", list(\"python_pkg_check\"), \"\\n\", list(\"\\n\", \"python_pkg_check(\\n\", \"  req.package,\\n\", \"  python.exec.path = NULL,\\n\", \"  python.lib.path = NULL,\\n\", \"  lib.path = NULL\\n\", \")\\n\"), \"\\n\", list(\"\\n\", list(list(\"req.package\"), list(\"Character vector of package names to import.\")), \"\\n\", \n    \"\\n\", list(list(\"python.exec.path\"), list(\"Single optional character vector specifying the absolute pathways of the executable python file to use (associated to the packages to use). If NULL, the reticulate::import_from_path() function used in python_pkg_check() seeks for an available version of python.exe, and then uses python_config(python_version, required_module, python_versions). But might not be the correct one for the python.lib.path parameter specified. Thus, it is recommanded to do not leave NULL, notably when using computing clusters.\")), \n    \"\\n\", \"\\n\", list(list(\"python.lib.path\"), list(\"Optional character vector specifying the absolute pathways of the directories containing some of the listed packages in the req.package argument, if not in the default directories.\")), \"\\n\", \"\\n\", list(list(\"lib.path\"), list(\"Absolute path of the reticulate packages, if not in the default folders.\")), \"\\n\"), \"\\n\", list(\"\\n\", \"Nothing.\\n\"), \"\\n\", list(\"\\n\", \"Check if the specified python packages are present in the computer (no import).\\n\"), \"\\n\", \n    list(\"\\n\", \"REQUIRED PACKAGES\\n\", \"\\n\", \"reticulate\\n\", \"\\n\", \"\\n\", \"REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION\\n\", \"\\n\", \"arg_check()\\n\", \"\\n\", \"WARNINGS\\n\", \"\\n\", \"for python 3.7. Previous versions return an error \\\"Error in sys$stdout$flush() : attempt to apply non-function\\\"\\n\"), \"\\n\", list(\"\\n\", \"# example of error message\\n\", \"\\n\", \"# python_pkg_check(req.package = \\\"nopackage\\\")\\n\", \"\\n\", \"\\n\", \"# commented because this example returns an error if the python package is not installed on the computer\\n\", \n        \"# (require the installation of the python serpentine package \\n\", \"# from https://github.com/koszullab/serpentine\\n\", \"\\n\", \"# python_pkg_check(req.package = \\\"serpentine\\\")\\n\", \"\\n\", \"\\n\", \"# another example of error message\\n\", \"\\n\", \"# python_pkg_check(req.package = \\\"serpentine\\\", python.lib.path = \\\"blablabla\\\")\\n\"), \"\\n\"), report.Rd = list(\"% Generated by roxygen2: do not edit by hand\", \"\\n\", \"% Please edit documentation in R/report.R\", \"\\n\", list(\"report\"), \"\\n\", list(\"report\"), \"\\n\", \n    list(\"report\"), \"\\n\", list(\"\\n\", \"report(\\n\", \"  data,\\n\", \"  output = \\\"results.txt\\\",\\n\", \"  path,\\n\", \"  overwrite = FALSE,\\n\", \"  rownames.kept = FALSE,\\n\", \"  vector.cat = FALSE,\\n\", \"  noquote = TRUE,\\n\", \"  sep = 2\\n\", \")\\n\"), \"\\n\", list(\"\\n\", list(list(\"data\"), list(\"Object to print in the output file. If NULL, nothing is done, with no warning.\")), \"\\n\", \"\\n\", list(list(\"output\"), list(\"Single character string. Name of the output file.\")), \"\\n\", \"\\n\", list(list(\"path\"), list(\"Location of the output file.\")), \n        \"\\n\", \"\\n\", list(list(\"overwrite\"), list(\"Single logical value. If output file already exists, defines if the printing is appended (default FALSE) or if the output file content is erased before printing (TRUE).\")), \"\\n\", \"\\n\", list(list(\"rownames.kept\"), list(\"Single logical value. Defines whether row names have to be removed or not in small tables (less than length.rows rows).\")), \"\\n\", \"\\n\", list(list(\"vector.cat\"), list(\"Single logical value. If TRUE print a vector of length > 1 using cat() instead of capture.output(). Otherwise (default FALSE) the opposite.\")), \n        \"\\n\", \"\\n\", list(list(\"noquote\"), list(\"Single logical value. If TRUE no quote are present for the characters.\")), \"\\n\", \"\\n\", list(list(\"sep\"), list(\"Single integer representing the number of empty lines after printed data.\")), \"\\n\"), \"\\n\", list(\"\\n\", \"Nothing.\\n\"), \"\\n\", list(\"\\n\", \"Log file function: print a character string or a data object into a same output file.\\n\"), \"\\n\", list(\"\\n\", \"REQUIRED PACKAGES\\n\", \"\\n\", \"none\\n\", \"\\n\", \"\\n\", \"REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION\\n\", \n        \"\\n\", \"arg_check()\\n\"), \"\\n\", list(\"\\n\", \"#report()\\n\", \"report(data = 1:3, output = \\\"results.txt\\\", path = \\\".\\\", overwrite = TRUE, \\n\", \"rownames.kept = FALSE, vector.cat = FALSE, noquote = FALSE, sep = 2)\\n\"), \"\\n\")), source = list(arg_check.Rd = \"R/arg_check.R\", arg_test.Rd = \"R/arg_test.R\", env_check.Rd = \"R/env_check.R\", get_message.Rd = \"R/get_message.R\", internal_function.Rd = \"R/internal_functions.R\", is_function_here.Rd = \"R/is_function_here.R\", is_package_here.Rd = \"R/is_package_here.R\", \n    python_pkg_check.Rd = \"R/python_pkg_check.R\", report.Rd = \"R/report.R\"), keywords = list(character(0), character(0), character(0), character(0), \"internal\", character(0), character(0), character(0), character(0)), concepts = list(character(0), character(0), character(0), character(0), character(0), character(0), character(0), character(0), character(0)), internal = c(FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, FALSE)), tutorials = list(name = character(0), file_out = character(0), \n    title = character(0), pagetitle = character(0), url = character(0)), vignettes = list(name = c(\"articles/arg_check\", \"articles/arg_test\", \"articles/env_check\", \"articles/get_message\", \"articles/is_function_here\", \"articles/is_package_here\", \"articles/python_pkg_check\", \"articles/report\", \"cuteDev\", \"exemples\"), file_in = c(\"vignettes/articles/arg_check.Rmd\", \"vignettes/articles/arg_test.Rmd\", \"vignettes/articles/env_check.Rmd\", \"vignettes/articles/get_message.Rmd\", \"vignettes/articles/is_function_here.Rmd\", \n\"vignettes/articles/is_package_here.Rmd\", \"vignettes/articles/python_pkg_check.Rmd\", \"vignettes/articles/report.Rmd\", \"vignettes/cuteDev.Rmd\", \"vignettes/exemples.Rmd\"), file_out = c(\"articles/arg_check.html\", \"articles/arg_test.html\", \"articles/env_check.html\", \"articles/get_message.html\", \"articles/is_function_here.html\", \"articles/is_package_here.html\", \"articles/python_pkg_check.html\", \"articles/report.html\", \"articles/cuteDev.html\", \"articles/exemples.html\"), title = c(\"arg_check\", \"arg_test\", \n\"env_check\", \"get_message\", \"is_function_here\", \"is_package_here\", \"python_pkg_check\", \"report\", \"cuteDev\", \"exemples\"), description = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA), depth = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1)), bs_version = 3, prefix = \"\"), examples = TRUE, run_dont_run = FALSE, seed = 1014, lazy = FALSE, override = list(), install = FALSE, preview = FALSE, new_process = FALSE, devel = FALSE, cli_colors = 256, pkgdown_internet = TRUE)), envir = base::quote(<environment>), quote = base::quote(TRUE)): envir\ndo.call(do.call, c(readRDS(\"C:\\\\Users\\\\yhan\\\\AppData\\\\Local\\\\Temp\\\\RtmpAdZd7c\\\\callr-fun-3938162049f2\"), list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv, quote = TRUE): envir\nwithCallingHandlers({\n    NULL\n    saveRDS(do.call(do.call, c(readRDS(\"C:\\\\Users\\\\yhan\\\\AppData\\\\Local\\\\Temp\\\\RtmpAdZd7c\\\\callr-fun-3938162049f2\"), list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = \"C:\\\\Users\\\\yhan\\\\AppData\\\\Local\\\\Temp\\\\RtmpAdZd7c\\\\callr-res-3938542bee3\", compress = FALSE)\n    flush(stdout())\n    flush(stderr())\n    NULL\n    invisible()\n}, error = function(e) {\n    {\n        callr_data <- as.environment(\"tools:callr\")$`__callr_data__`\n        err <- callr_data$err\n        if (FALSE) {\n            assign(\".Traceback\", .traceback(4), envir = callr_data)\n            dump.frames(\"__callr_dump__\")\n            assign(\".Last.dump\", .GlobalEnv$`__callr_dump__`, envir = callr_data)\n            rm(\"__callr_dump__\", envir = .GlobalEnv)\n        }\n        e <- err$process_call(e)\n        e2 <- err$new_error(\"error in callr subprocess\")\n        class(e2) <- c(\"callr_remote_error\", class(e2))\n        e2 <- err$add_trace_back(e2)\n        cut <- which(e2$trace$scope == \"global\")[1]\n        if (!is.na(cut)) {\n            e2$trace <- e2$trace[-(1:cut), ]\n        }\n        saveRDS(list(\"error\", e2, e), file = paste0(\"C:\\\\Users\\\\yhan\\\\AppData\\\\Local\\\\Temp\\\\RtmpAdZd7c\\\\callr-res-3938542bee3\", \".error\"))\n    }\n}, interrupt = function(e) {\n    {\n        callr_data <- as.environment(\"tools:callr\")$`__callr_data__`\n        err <- callr_data$err\n        if (FALSE) {\n            assign(\".Traceback\", .traceback(4), envir = callr_data)\n            dump.frames(\"__callr_dump__\")\n            assign(\".Last.dump\", .GlobalEnv$`__callr_dump__`, envir = callr_data)\n            rm(\"__callr_dump__\", envir = .GlobalEnv)\n        }\n        e <- err$process_call(e)\n        e2 <- err$new_error(\"error in callr subprocess\")\n        class(e2) <- c(\"callr_remote_error\", class(e2))\n        e2 <- err$add_trace_back(e2)\n        cut <- which(e2$trace$scope == \"global\")[1]\n        if (!is.na(cut)) {\n            e2$trace <- e2$trace[-(1:cut), ]\n        }\n        saveRDS(list(\"error\", e2, e), file = paste0(\"C:\\\\Users\\\\yhan\\\\AppData\\\\Local\\\\Temp\\\\RtmpAdZd7c\\\\callr-res-3938542bee3\", \".error\"))\n    }\n}, callr_message = function(e) {\n    try(signalCondition(e))\n}): expr\ndoTryCatch(return(expr), name, parentenv, handler): expr\ntryCatchOne(expr, names, parentenv, handlers[[1]]): expr\ntryCatchList(expr, names[-nh], parentenv, handlers[-nh]): expr\ndoTryCatch(return(expr), name, parentenv, handler): expr\ntryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]), names[nh], parentenv, handlers[[nh]]): expr\ntryCatchList(expr, classes, parentenv, handlers): expr\ntryCatch(withCallingHandlers({\n    NULL\n    saveRDS(do.call(do.call, c(readRDS(\"C:\\\\Users\\\\yhan\\\\AppData\\\\Local\\\\Temp\\\\RtmpAdZd7c\\\\callr-fun-3938162049f2\"), list(envir = .GlobalEnv, quote = TRUE)), envir = .GlobalEnv, quote = TRUE), file = \"C:\\\\Users\\\\yhan\\\\AppData\\\\Local\\\\Temp\\\\RtmpAdZd7c\\\\callr-res-3938542bee3\", compress = FALSE)\n    flush(stdout())\n    flush(stderr())\n    NULL\n    invisible()\n}, error = function(e) {\n    {\n        callr_data <- as.environment(\"tools:callr\")$`__callr_data__`\n        err <- callr_data$err\n        if (FALSE) {\n            assign(\".Traceback\", .traceback(4), envir = callr_data)\n            dump.frames(\"__callr_dump__\")\n            assign(\".Last.dump\", .GlobalEnv$`__callr_dump__`, envir = callr_data)\n            rm(\"__callr_dump__\", envir = .GlobalEnv)\n        }\n        e <- err$process_call(e)\n        e2 <- err$new_error(\"error in callr subprocess\")\n        class(e2) <- c(\"callr_remote_error\", class(e2))\n        e2 <- err$add_trace_back(e2)\n        cut <- which(e2$trace$scope == \"global\")[1]\n        if (!is.na(cut)) {\n            e2$trace <- e2$trace[-(1:cut), ]\n        }\n        saveRDS(list(\"error\", e2, e), file = paste0(\"C:\\\\Users\\\\yhan\\\\AppData\\\\Local\\\\Temp\\\\RtmpAdZd7c\\\\callr-res-3938542bee3\", \".error\"))\n    }\n}, interrupt = function(e) {\n    {\n        callr_data <- as.environment(\"tools:callr\")$`__callr_data__`\n        err <- callr_data$err\n        if (FALSE) {\n            assign(\".Traceback\", .traceback(4), envir = callr_data)\n            dump.frames(\"__callr_dump__\")\n            assign(\".Last.dump\", .GlobalEnv$`__callr_dump__`, envir = callr_data)\n            rm(\"__callr_dump__\", envir = .GlobalEnv)\n        }\n        e <- err$process_call(e)\n        e2 <- err$new_error(\"error in callr subprocess\")\n        class(e2) <- c(\"callr_remote_error\", class(e2))\n        e2 <- err$add_trace_back(e2)\n        cut <- which(e2$trace$scope == \"global\")[1]\n        if (!is.na(cut)) {\n            e2$trace <- e2$trace[-(1:cut), ]\n        }\n        saveRDS(list(\"error\", e2, e), file = paste0(\"C:\\\\Users\\\\yhan\\\\AppData\\\\Local\\\\Temp\\\\RtmpAdZd7c\\\\callr-res-3938542bee3\", \".error\"))\n    }\n}, callr_message = function(e) {\n    try(signalCondition(e))\n}), error = function(e) {\n    NULL\n    if (FALSE) {\n        try(stop(e))\n    }\n    else {\n        invisible()\n    }\n}, interrupt = function(e) {\n    NULL\n    if (FALSE) {\n        e\n    }\n    else {\n        invisible()\n    }\n}): expr\n"
# test if any object of the stats environment (one step above .GlobalEnv) 
# are above environments. Returns NULL since no object names of stats are in above environments


# Example inside a function

fun1 <- function(){t.test <- 0 ; mean <- 5 ; env_check(pos = 1)} 
# env_check() will check if the object names inside the fun1 function 
# exist in the .GlobalEnv environment and above.
fun1()
#> [1] "SOME VARIABLES OF THE CHECKED ENVIRONMENT ARE ALSO PRESENT IN :\neval(expr, envir, enclos): mean t.test\npackage:stats: t.test\npackage:base: mean\n"
fun2 <- function(){t.test <- 0 ; mean <- 5 ; env_check(pos = 2)} 
# env_check() will check if the object names inside the fun2 function 
# exist in the stats environment and above.
fun2()
#> [1] "SOME VARIABLES OF THE CHECKED ENVIRONMENT ARE ALSO PRESENT IN :\npackage:stats: t.test\npackage:base: mean\n"
fun3 <- function(){t.test <- 0 ; mean <- 5 ; env_check(pos = 2, name = "fun3")} 
# idem fun2() but with the name of the function fun2 indicated. 
# Instead of writting name = "fun3", 
# we can also use name = as.character(sys.calls()[[length(sys.calls())]]), 
# as sys.calls() gives the function name at top stack of the imbricated functions, 
# sys.calls()[[length(sys.calls())]] the name of the just above function. 
# This can also been used for the above function: as.character(sys.call(1))
fun3()
#> [1] "SOME VARIABLES OF fun3 ARE ALSO PRESENT IN :\npackage:stats: t.test\npackage:base: mean\n"
test.pos <- 1
env_check(pos = test.pos, 
name = if(length(sys.calls()) >= test.pos)
{as.character(sys.calls()[[length(sys.calls()) + 1 - test.pos]])}
else{search()[ (1:length(search()))[test.pos - length(sys.calls())]]}) 
#> [1] "SOME VARIABLES OF eval ARE ALSO PRESENT IN :\npackage:stats: t.test\npackage:base: mean\n"
# here is a way to have the name of the tested environment according to test.pos value