get_message

Evaluate an instruction written between "" and return the first of the error, or warning or standard (non error non warning) messages if ever exist.

Using argument print.no = FALSE, return NULL if no message, which is convenient in some cases.

get_message(
  data,
  kind = "error",
  header = TRUE,
  print.no = FALSE,
  text = NULL,
  env = NULL
)

Arguments

data

Single character string to evaluate.

kind

Single character string. Either "error" to get error messages, or "warning" to get warning messages, or "message" to get non error and non warning messages.

header

Single logical value. Add a header in the returned message?

print.no

Single logical value. Print a message saying that no message reported?

text

Single character string added to the output message (even if no message exists and print.no is TRUE). Inactivated if the header argument is FALSE. Write NULL if not required.

env

The name of an existing environment. Write NULL if not required.

Value

The message or NULL if no message and print.no is FALSE.

Details

REQUIRED PACKAGES

ggplot2

REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

arg_check()

pkg_check()

WARNINGS

Only the first message is returned.

Always use the env argument when get_message() is used inside functions.

The function does not prevent printing if print() is used inside the instruction tested. To prevent that, use tempo <- utils::capture.output(error <- get_message(data = "arg_check(data = 'a', class = mean, neg.values = FALSE, print = TRUE)")). The return of get_message() is assigned into error and the printed messages are captured by utils::capture.output() and assigned into tempo. See the examples.

Examples

get_message(data = "wilcox.test(c(1,1,3), c(1, 2, 4), paired = TRUE)", kind = "error", 
print.no = TRUE, text = "IN A")
#> [1] "NO ERROR MESSAGE REPORTED IN A"

get_message(data = "wilcox.test(c(1,1,3), c(1, 2, 4), paired = TRUE)", kind = "warning", 
print.no = TRUE, text = "IN A")
#> [1] "WARNING MESSAGE REPORTED IN A:\nsimpleWarning in wilcox.test.default(c(1, 1, 3), c(1, 2, 4), paired = TRUE): cannot compute exact p-value with zeroes\n"

get_message(data = "wilcox.test(c(1,1,3), c(1, 2, 4), paired = TRUE)", kind = "message", 
print.no = TRUE, text = "IN A")
#> [1] "NO STANDARD (NON ERROR AND NON WARNING) MESSAGE REPORTED IN A"

get_message(data = "wilcox.test()", kind = "error", print.no = TRUE, text = "IN A")
#> [1] "ERROR MESSAGE REPORTED IN A:\nIn wilcox.test.default() : argument \"x\" is missing, with no default\n"

get_message(data = "sum(1)", kind = "error", print.no = TRUE, text = "IN A")
#> [1] "NO ERROR MESSAGE REPORTED IN A"

get_message(data = "message('ahah')", kind = "error", print.no = TRUE, text = "IN A")
#> [1] "NO ERROR MESSAGE REPORTED IN A"

get_message(data = "message('ahah')", kind = "message", print.no = TRUE, text = "IN A")
#> ahah
#> [1] "NO STANDARD (NON ERROR AND NON WARNING) MESSAGE REPORTED IN A"

get_message(data = "ggplot(data = data.frame(X = 1:10, stringsAsFactors = TRUE), 
mapping = aes(x = X)) + geom_histogram()", kind = "message", print.no = TRUE, 
text = "IN FUNCTION 1")
#> `stat_bin()` using `bins = 30`. Pick better value with `binwidth`.
#> [1] "NO STANDARD (NON ERROR AND NON WARNING) MESSAGE REPORTED IN FUNCTION 1"

set.seed(1) ; 
obs1 <- data.frame(Time = c(rnorm(10), rnorm(10) + 2), 
Group1 = rep(c("G", "H"), each = 10), stringsAsFactors = TRUE) ; 
get_message(data = 'gg_boxplot(data = obs1, y = "Time", categ = "Group1")', 
kind = "message", print.no = TRUE, text = "IN FUNCTION 1")
#> [1] "NO STANDARD (NON ERROR AND NON WARNING) MESSAGE BECAUSE OF ERROR MESSAGE REPORTED IN FUNCTION 1"