report

Log file function: print a character string or a data object into a same output file.

report(
  data,
  output = "results.txt",
  path,
  overwrite = FALSE,
  rownames.kept = FALSE,
  vector.cat = FALSE,
  noquote = TRUE,
  sep = 2
)

Arguments

data

Object to print in the output file. If NULL, nothing is done, with no warning.

output

Single character string. Name of the output file.

path

Location of the output file.

overwrite

Single logical value. If output file already exists, defines if the printing is appended (default FALSE) or if the output file content is erased before printing (TRUE).

rownames.kept

Single logical value. Defines whether row names have to be removed or not in small tables (less than length.rows rows).

vector.cat

Single logical value. If TRUE print a vector of length > 1 using cat() instead of capture.output(). Otherwise (default FALSE) the opposite.

noquote

Single logical value. If TRUE no quote are present for the characters.

sep

Single integer representing the number of empty lines after printed data.

Value

Nothing.

Details

REQUIRED PACKAGES

none

REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

arg_check()

Examples

#report()
report(data = 1:3, output = "results.txt", path = ".", overwrite = TRUE, 
rownames.kept = FALSE, vector.cat = FALSE, noquote = FALSE, sep = 2)