env_check

Verify that object names in the environment defined by the pos parameter are identical or not to object names in the above environment (following R Scope). This can be used to verify that names used for objects inside a function or in the working environment do not override names of objects already present in the above R environments, following the R scope.

env_check(pos = 1, name = NULL)

Arguments

pos

Single non nul positive integer indicating the position of the environment checked (argument n of the parent.frame() function). Value 1 means one step above the env_check() local environment (by default). This means that when env_check(pos = 1) is used inside a function A, it checks if the name of any object in the local environment of this function A is also present in above environments, following R Scope, starting by the just above environment. When env_check(pos = 1) is used in the working (Global) environment (named .GlobalEnv), it checks the object names of this .GlobalEnv environment, in the above environments. See the examples below.

name

Single character string indicating a string that will be added in the output string, for instance the name of a function inside which env_check() is used.

Value

A character string indicating the object names of the tested environment that match object names in the above environments, following the R scope, or NULL if no match.

Details

REQUIRED PACKAGES

None

REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

arg_check()

Examples

# Example in the working environment

mean <- 1 
# creation of the object mean with value 1 in the .GlobalEnv environment, 
# knowing that the mean() function also exists in the environment base, above .GlobalEnv.
t.test <- 1 
# creation of the object t.test with value 1 in the .GlobalEnv environment, 
# knowing that the t.test() function also exists in the environment stats, above .GlobalEnv.
search() 
#>  [1] ".GlobalEnv"        "package:cuteDev"   "package:stats"    
#>  [4] "package:graphics"  "package:grDevices" "package:utils"    
#>  [7] "package:datasets"  "package:methods"   "Autoloads"        
#> [10] "package:base"     
# current R scope (order of the successive R environments).
utils::find("mean") 
#> [1] "package:base"
# where the objects with the name "mean" are present.
utils::find("t.test") 
#> [1] "package:stats"
# where the objects with the name "mean" are present.
a <- env_check(pos = 1) 
# test if any object name of the global environment are above environments 
# (or env_check(), as pos = 1 is the default value).
a # the output string of sec().
#> [1] "SOME VARIABLES OF THE CHECKED ENVIRONMENT ARE ALSO PRESENT IN :\npackage:stats: t.test\npackage:base: mean\n"
cat(a) # the evaluated output.
#> SOME VARIABLES OF THE CHECKED ENVIRONMENT ARE ALSO PRESENT IN :
#> package:stats: t.test
#> package:base: mean
env_check(pos = 2) 
#> [1] "SOME VARIABLES OF THE CHECKED ENVIRONMENT ARE ALSO PRESENT IN :\neval_with_user_handlers(expr, envir, enclos, user_handlers): enclos envir expr\nwithCallingHandlers(withVisible(eval_with_user_handlers(expr, envir, enclos, user_handlers)), warning = wHandler, error = eHandler, message = mHandler): expr\ndoTryCatch(return(expr), name, parentenv, handler): expr\ntryCatchOne(expr, names, parentenv, handlers[[1]]): expr\ntryCatchList(expr, classes, parentenv, handlers): expr\ntryCatch(expr, error = function(e) {\n    call <- conditionCall(e)\n    if (!is.null(call)) {\n        if (identical(call[[1]], quote(doTryCatch))) \n            call <- sys.call(-4)\n        dcall <- deparse(call, nlines = 1)\n        prefix <- paste(\"Error in\", dcall, \": \")\n        LONG <- 75\n        sm <- strsplit(conditionMessage(e), \"\\n\")[[1]]\n        w <- 14 + nchar(dcall, type = \"w\") + nchar(sm[1], type = \"w\")\n        if (is.na(w)) \n            w <- 14 + nchar(dcall, type = \"b\") + nchar(sm[1], type = \"b\")\n        if (w > LONG) \n            prefix <- paste0(prefix, \"\\n  \")\n    }\n    else prefix <- \"Error : \"\n    msg <- paste0(prefix, conditionMessage(e), \"\\n\")\n    .Internal(seterrmessage(msg[1]))\n    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {\n        cat(msg, file = outFile)\n        .Internal(printDeferredWarnings())\n    }\n    invisible(structure(msg, class = \"try-error\", condition = e))\n}): expr\ntry(f, silent = TRUE): expr\nevaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, debug = debug, last = i == length(out), use_try = stop_on_error != 2, keep_warning = keep_warning, keep_message = keep_message, log_echo = log_echo, log_warning = log_warning, output_handler = output_handler, include_timing = include_timing): enclos envir expr\nevaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler): enclos envir expr\nwithCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env, run_dont_run = run_dont_run), error = function(err) {\n    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), i = err$message)\n    abort(msg, parent = err)\n}): expr\nwithCallingHandlers(expr, error = function(cnd) {\n    if (i == 0) {\n    }\n    else {\n        message <- c(i = \"In index: {i}.\")\n        if (!is.null(names) && !is.na(names[[i]]) && names[[i]] != \"\") {\n            name <- names[[i]]\n            message <- c(message, i = \"With name: {name}.\")\n        }\n        else {\n            name <- NULL\n        }\n        cli::cli_abort(message, location = i, name = name, parent = cnd, call = error_call, class = \"purrr_error_indexed\")\n    }\n}): expr\nwith_indexed_errors(i = i, names = names, error_call = .purrr_error_call, call_with_cleanup(map_impl, environment(), .type, .progress, n, names, i)): expr\n"
# test if any object of the stats environment (one step above .GlobalEnv) 
# are above environments. Returns NULL since no object names of stats are in above environments

# Example inside a function

fun1 <- function(){t.test <- 0 ; mean <- 5 ; env_check(pos = 1)} 
# env_check() will check if the object names inside the fun1 function 
# exist in the .GlobalEnv environment and above.
fun1()
#> [1] "SOME VARIABLES OF THE CHECKED ENVIRONMENT ARE ALSO PRESENT IN :\neval(expr, envir, enclos): mean t.test\npackage:stats: t.test\npackage:base: mean\n"
fun2 <- function(){t.test <- 0 ; mean <- 5 ; env_check(pos = 2)} 
# env_check() will check if the object names inside the fun2 function 
# exist in the stats environment and above.
fun2()
#> [1] "SOME VARIABLES OF THE CHECKED ENVIRONMENT ARE ALSO PRESENT IN :\npackage:stats: t.test\npackage:base: mean\n"
fun3 <- function(){t.test <- 0 ; mean <- 5 ; env_check(pos = 2, name = "fun3")} 
# idem fun2() but with the name of the function fun2 indicated. 
# Instead of writting name = "fun3", 
# we can also use name = as.character(sys.calls()[[length(sys.calls())]]), 
# as sys.calls() gives the function name at top stack of the imbricated functions, 
# sys.calls()[[length(sys.calls())]] the name of the just above function. 
# This can also been used for the above function: as.character(sys.call(1))
fun3()
#> [1] "SOME VARIABLES OF fun3 ARE ALSO PRESENT IN :\npackage:stats: t.test\npackage:base: mean\n"
test.pos <- 1
env_check(pos = test.pos, 
name = if(length(sys.calls()) >= test.pos)
{as.character(sys.calls()[[length(sys.calls()) + 1 - test.pos]])}
else{search()[ (1:length(search()))[test.pos - length(sys.calls())]]}) 
#> [1] "SOME VARIABLES OF eval ARE ALSO PRESENT IN :\npackage:stats: t.test\npackage:base: mean\n"
# here is a way to have the name of the tested environment according to test.pos value